+++ credentials/src/index.ts
@@ -1,4 +1,3 @@
 export * from "./jubjub.js";
 export type { OffchainMidnightHolderBinding as OffchainDIDHolderBinding } from "./managed/credentials/contract/index.js";
-export * as CredentialsContract from "./managed/credentials/contract/index.js";
 export * from "./managed/credentials/contract/index.js";